Implementation for #200
This is a suggestion for the implementation of #200

I already implemented exactly this code at a customer. It introduces two new variables to var.tfvars to control how many logical ports shall be used for VNIC failover and also to define the capacity of the logical ports.

Signed-off-by: Torsten Wendland <twendlan@de.ibm.com>

PS:
Added validation of number of VFs, updated documentation with min and max, changed variable name from sriov_vnic_failover_adapter to sriov_vnic_failover_vfs (because we don't define amount of adapters, we are defining amount of virtual functions; adapters might mislead the user)